+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 * Added Gender Filter to Team Overview in Django Admin
 
 ### Fixes
+* Fixed several Translation issues in Django Admin
 
 ### Breaks